break up song middleware
