+++ src/store/middleware/audio.middleware.ts
@@ -1,30 +1,15 @@
 import { createListenerMiddleware } from "@reduxjs/toolkit";
-import { ActionCreators as ReduxUndoActionCreators } from "redux-undo";
 
-import { HIGHEST_PRECISION } from "$/constants";
-import { convertBeatsToMilliseconds, convertMillisecondsToBeats, getWaveformDataForFile, snapToNearestBeat } from "$/helpers/audio.helpers";
-import { convertBookmarksToRedux } from "$/helpers/bookmarks.helpers";
-import { convertEventsToExportableJson, convertEventsToRedux } from "$/helpers/events.helpers";
+import { convertBeatsToMilliseconds, convertMillisecondsToBeats, snapToNearestBeat } from "$/helpers/audio.helpers";
 import { convertFileToArrayBuffer } from "$/helpers/file.helpers";
-import { convertNotesFromMappingExtensions } from "$/helpers/notes.helpers";
-import { convertObstaclesToRedux } from "$/helpers/obstacles.helpers";
 import { AudioSample } from "$/services/audio.service";
-import { FileType, deleteAllSongFiles, deleteFile, getBeatmap, getFile, getFilenameForThing, saveBeatmap, saveFile, saveInfoDat } from "$/services/file.service";
-import { createBeatmapContents, createInfoContent } from "$/services/packaging.service";
-import { shiftEntitiesByOffset, unshiftEntitiesByOffset } from "$/services/packaging.service.nitty-gritty";
+import { getFile } from "$/services/file.service";
 import { Sfx } from "$/services/sfx.service";
 import {
 	adjustCursorPosition,
-	copyDifficulty,
-	createDifficulty,
-	deleteBeatmap,
-	deleteSong,
-	finishLoadingSong,
 	jumpToBeat,
 	leaveEditor,
-	loadBeatmapEntities,
 	pausePlaying,
-	reloadWaveform,
 	scrollThroughSong,
 	scrubEventsHeader,
 	scrubWaveform,
@@ -42,10 +27,10 @@ import {
 	updateSongDetails,
 	updateVolume,
 } from "$/store/actions";
-import { getAllEventsAsArray, getBeatsPerZoomLevel, getCursorPosition, getCursorPositionInBeats, getDuration, getIsLockedToCurrentWindow, getNotes, getPlayNoteTick, getPlaybackRate, getProcessingDelay, getSelectedSong, getSnapTo, getSongById, getVolume } from "$/store/selectors";
+import { getBeatsPerZoomLevel, getCursorPosition, getCursorPositionInBeats, getDuration, getIsLockedToCurrentWindow, getNotes, getPlayNoteTick, getPlaybackRate, getProcessingDelay, getSelectedSong, getSnapTo, getSongById, getVolume } from "$/store/selectors";
 import type { RootState } from "$/store/setup";
-import { type Json, View } from "$/types";
-import { clamp, floorToNearest, roundToNearest } from "$/utils";
+import { View } from "$/types";
+import { clamp, floorToNearest } from "$/utils";
 
 function stopAndRewindAudio(audioSample: AudioSample, offset: number) {
 	audioSample.setCurrentTime((offset || 0) / 1000);
@@ -82,10 +67,8 @@ function calculateIfPlaybackShouldBeCommandeered(state: RootState, currentBeat: 
 	}
 }
 
-/**
- * This middleware manages playback concerns.
- */
-export default function createSongMiddleware() {
+/** This middleware manages playback concerns. */
+export default function createAudioMiddleware() {
 	let animationFrameId: number;
 
 	const ticker = new Sfx();
@@ -97,7 +80,7 @@ export default function createSongMiddleware() {
 		actionCreator: startLoadingSong,
 		effect: async (action, api) => {
 			api.unsubscribe();
-			const { songId, difficulty } = action.payload;
+			const { songId } = action.payload;
 			const state = api.getState();
 			const song = getSongById(state, songId);
 			const volume = getVolume(state);
@@ -106,38 +89,6 @@ export default function createSongMiddleware() {
 				console.error(`Song "${songId}" not found. Current state:`, state);
 				return;
 			}
-			// Fetch the json for this beatmap from our local store.
-			let beatmapJson: Json.Beatmap | null = null;
-			try {
-				beatmapJson = await getBeatmap(songId, difficulty);
-			} catch (err) {
-				console.error(err);
-			}
-			// we may not have any beatmap entities, if this is a new song or new difficulty.
-			if (beatmapJson) {
-				let notes = beatmapJson._notes;
-				// If this song uses mapping extensions, the note values will be in the thousands. We need to pull them down to the normal range.
-				if (song.modSettings.mappingExtensions?.isEnabled) {
-					notes = convertNotesFromMappingExtensions(notes);
-				}
-				// If we do, we need to manage a little dance related to offsets.
-				// See offsets.md for more context, but essentially we need to transform our timing to match the beat, by undoing a transformation previously applied.
-				let unshiftedNotes = unshiftEntitiesByOffset(notes || [], song.offset, song.bpm);
-				const unshiftedEvents = unshiftEntitiesByOffset(beatmapJson._events || [], song.offset, song.bpm);
-				const unshiftedObstacles = unshiftEntitiesByOffset(beatmapJson._obstacles || [], song.offset, song.bpm);
-				// Round all notes, so that no floating-point imprecision drift happens
-				unshiftedNotes = unshiftedNotes.map((note) => {
-					return { ...note, _time: roundToNearest(note._time, HIGHEST_PRECISION) };
-				});
-				// Our beatmap comes in a "raw" form, using proprietary fields.
-				// At present, I'm using that proprietary structure for notes/mines, but I have my own structure for obstacles and events.
-				// So I need to convert the ugly JSON format to something manageable.
-				const convertedObstacles = convertObstaclesToRedux(unshiftedObstacles as Json.Obstacle[]);
-				const convertedEvents = convertEventsToRedux(unshiftedEvents as Json.Event[]);
-				const convertedBookmarks = beatmapJson._customData?._bookmarks ? convertBookmarksToRedux(beatmapJson._customData._bookmarks) : [];
-				api.dispatch(loadBeatmapEntities({ notes: unshiftedNotes as Json.Note[], events: convertedEvents, obstacles: convertedObstacles, bookmarks: convertedBookmarks }));
-				api.dispatch(ReduxUndoActionCreators.clearHistory());
-			}
 			const file = await getFile<Blob>(song.songFilename);
 			if (!file) return;
 			const arrayBuffer = await convertFileToArrayBuffer(file);
@@ -145,44 +96,7 @@ export default function createSongMiddleware() {
 			audioSample.changeVolume(volume);
 			audioSample.changePlaybackRate(playbackRate);
 			audioSample.setCurrentTime(song.offset / 1000);
-			const waveform = await getWaveformDataForFile(file);
-			api.dispatch(finishLoadingSong({ song, waveformData: waveform }));
 			api.subscribe();
-		},
-	});
-	instance.startListening({
-		actionCreator: createDifficulty,
-		effect: async (action, api) => {
-			const { difficulty, afterCreate } = action.payload;
-			const state = api.getState();
-			const song = getSelectedSong(state);
-			const events = convertEventsToExportableJson(getAllEventsAsArray(state));
-			const shiftedEvents = shiftEntitiesByOffset(events, song.offset, song.bpm);
-			// No notes/obstacles/bookmarks by default, but copy the lighting
-			const beatmapContents = createBeatmapContents({ notes: [], obstacles: [], events: shiftedEvents, bookmarks: [] }, { version: 2 });
-			const beatmapFilename = getFilenameForThing(song.id, FileType.BEATMAP, { difficulty });
-			await saveFile(beatmapFilename, beatmapContents);
-			if (typeof afterCreate === "function") {
-				afterCreate(difficulty);
-			}
-		},
-	});
-	instance.startListening({
-		actionCreator: copyDifficulty,
-		effect: async (action, api) => {
-			const state = api.getState();
-			const { songId, fromDifficultyId, toDifficultyId, afterCopy } = action.payload;
-			// First, we need to load the file which contains the notes, events, etc for the difficulty we want to copy.
-			const sourceDifficultyFileContents = await getBeatmap(songId, fromDifficultyId);
-			// Save it to our destination difficulty.
-			await saveBeatmap(songId, toDifficultyId, JSON.stringify(sourceDifficultyFileContents));
-			// Pull that updated redux state and save it to our Info.dat
-			const song = getSongById(state, songId);
-			// Back up our latest data!
-			await saveInfoDat(song.id, createInfoContent(song, { version: 2 }));
-			if (typeof afterCopy === "function") {
-				afterCopy(toDifficultyId);
-			}
 		},
 	});
 	instance.startListening({
@@ -256,15 +170,12 @@ export default function createSongMiddleware() {
 		effect: async (action, api) => {
 			api.unsubscribe();
 			const { songFilename, offset } = action.payload;
-			// It's possible we updated the song file. We should reload it, so that the audio is properly updated.
 			if (songFilename) {
 				const file = await getFile<Blob>(songFilename);
 				if (!file) return;
 				const arrayBuffer = await convertFileToArrayBuffer(file);
 				await audioSample.load(arrayBuffer);
 				audioSample.setCurrentTime((offset ?? 0) / 1000);
-				const waveform = await getWaveformDataForFile(file);
-				api.dispatch(reloadWaveform({ waveformData: waveform }));
 			}
 			api.subscribe();
 		},
@@ -456,24 +367,6 @@ export default function createSongMiddleware() {
 			const { playbackRate } = action.payload;
 			audioSample.changePlaybackRate(playbackRate);
 			api.subscribe();
-		},
-	});
-	instance.startListening({
-		actionCreator: deleteBeatmap,
-		effect: async (action) => {
-			const { songId, difficulty } = action.payload;
-			// Our reducer will handle the redux state part, but we also need to delete the corresponding beatmap from the filesystem.
-			const beatmapFilename = getFilenameForThing(songId, FileType.BEATMAP, { difficulty });
-			await deleteFile(beatmapFilename);
-		},
-	});
-	instance.startListening({
-		actionCreator: deleteSong,
-		effect: (action, api) => {
-			const state = api.getState();
-			const { songId } = action.payload;
-			const song = getSongById(state, songId);
-			deleteAllSongFiles(song);
 		},
 	});
